requirements.txt
asgiref==3.2.7
auth==0.5.3
blinker==1.4
certifi==2020.4.5.1
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.6
django-cors-headers==3.3.0
djangorestframework==3.11.0
dnspython==1.16.0
eventlet==0.25.2
falcon==2.0.0
greenlet==0.4.15
gunicorn==20.0.4
idna==2.9
mongoengine==0.20.0
monotonic==1.5
pymongo==3.10.1
pytz==2020.1
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
Werkzeug==1.0.1
whitenoise==5.1.0
psycopg2==2.7.5
psycopg2-binary==2.8.5
django-heroku==0.3.1